requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.8
beautifulsoup4==4.12.3
python-dotenv==1.0.1
dnspython==2.6.1
python-nmap==0.7.1
jinja2==3.1.4
python-multipart
